Add post-request filtering to Transactions and Order Refunds
There was not a query param for these streams to ask for records after a
certain date. This caused a failure in the start date test because we
would get too many records.

This change also includes a call to update and write a state message. The
bookmarks test was failing because there wouldn't be an entry in the final
state to assert on.