Get tests passing on all streams (#97)
* Actually paginate the transactions stream

* WIP: Update pagination test to test all streams

* Break up streams between two shopify stores

* Fix test_pagination

- Restrict a test run to just the streams passed in as a parameter
- Don't create connection in the test

* Fix test_bookmarks

- Run the test twice, once with each store
- Restrict streams tested to what was passed as a param
- Don't create a connection in the test

* Fix test_start_date

- Run with only the new store
- Run with all streams

* Add task to remove config parameter

* Add error handling to Transactions requests

* Add post-request filtering to Transactions and Order Refunds

There was not a query param for these streams to ask for records after a
certain date. This caused a failure in the start date test because we
would get too many records.

This change also includes a call to update and write a state message. The
bookmarks test was failing because there wouldn't be an entry in the final
state to assert on.

* Wrap each store's test in a subtest

* Fix test_automatic

- Run test once per store
- Remove assertion on foreign key metadata

* Make pylint happy

* Fix whitespace

* Rename environment variables

* Trigger tests

* PR Feedback: Add bug comment in more places, API_LIMIT falls back to
config value

* Pin pylint to the previous working version

We saw a failure on `pylint-2.8.1`

```
tap_shopify/streams/base.py:146:12: R1730: Consider using 'updated_at_max = min(updated_at_max, stop_time)' instead of unnecessary if block (consider-using-min-builtin)
```