Introduce support for end_date parameter
This is intended to be used to pull back
date ranges that have historically not
been ingested, without needing to sync
all data. For example, if you are missing
data from 2023-01-01 to 2023-02-01,
you could set the start and end dates
to these values to retrieve just that
month of data, rather than potentially
syncing an extra 9 months of data.
This can also be used to incrementally
test your ELT workflows to ensure they
are compatible with older data schemas.